Added new Corrections file.  Added new branch to the channel class Now it has Corrected cubic top/bottom time of flight.  Is using the Cubic Fit times.
LendaChannel now holds the corrected cubic fit time of flight.

Lendabar uses the corrected cubic fit time of flight to make a
a top/bottom corrected cubic TOF variable.

LendaPacker now sets the corrected cubic time of flight for the
lendachannel in the ddaschannel2LendaChannel method

Changed build data script.  Now looks to see if there is a root
file with "RAW" in it.  If so it calls Raw2Cal instead of Evt2Cal

Added some new corrrections files for the jitters in runs 397,398