Fixed the small bug in the main Evt2Cal unpacking loop. Before it was while (! feof(file) && !siganl_recived)
the EOF bit is not set until you reach the end of file and then try
to read once more.  So now it is
while(!signal_recived){
	read_total_size
if (feof(file)){
	break;
}

This is probably the source of the errors that were seen with multiple
segment runs.  When there was one file the last bit of the file was
the change of state end run information which is ignored so it being
read twice isn't a problem.  With two segments the last bit of the file
can be a physics event.  So if it tries to read again it will still
think it is a physics event and try to processs it and crash