The implementation of the changes discussed in previous commit. The packing is slightly more modular to allow the repacking methods to call the same bits of code when repacking.  There is now ReMakeLendaEvent which will take the LendaChannels in the Bars and TheObjecScintillators and Call the RePackChannel Method on them.
At the Momement RePackChannel and will Get the map info for the channel,
recalculate the waveform analysis routines then call the same PackCalculatedValues
method that the intial packing calls.  It is doing this on the inEvent data.

ReMakeLendaEvent

Recalculates the values for the calculated values in the inEvent.  Copies
each bar over into a temporary map (Just as the initial packing does).  Loops
over the objectscintilators and copies them over to outEvent
and then Calls RePackChannel. Stores the ReCalculated Reference times in the
same type of structure as the initial packing.  Then calls FillReferenceTimesInEvent
where the bars are copied over from the temporary map to the outEvent structure.

NOTE: Currently ignores any channels in the UnMappedChannels vector.
NOTE: Calling the method will change values in inEvent