Added the better GetEnergy method that was on the OPT branch. Instead of integrating the whole pulse, it takes only a window around the maximum value.
Renamed all the get*** Methods in the Filter to Get***.  GetPulseHeight
methods now take a second argument to save the max height spot for
the new GetEnergy.

I think I have solved the problem with the short and long gate filters.
Thought there was a difference between the RePack branch and the Master
branch, but there was the same bug in both.  The Long/Short gates would
start a few clock tics away from the softwareCFD value.  When the Get
Zero crossing method would fail it would set softwareCFD to -10008
Then the Long/Short gates would start at -10000ish value and add
up gibberish.

Now the Long/Short Gates start at MaxSpot-4.  This way it won't
can't fail to find a start point and so it won't add up gibberish