LendaBar now has a BarAngle parameter.
LendaPacker now has the infastructure for setting the bar angle in the
event building.  The MapInfo class has a bar angle parameter.  The Make
LendaEvent method sets the bar angle in a similiar way that the bar
name is set.  Don't think that it will set the bar angle in remake
lendaevent.  Need to Check that at one point.

Also The packer now has a method to get the map info object for a
given channel.

Changed the GetZeroCrossing calls to GetZeroCrossingImproved calls.
See next commit for the difference