Cal2Cal has been updated slightly.  It is smarter about which corrections files to use like Evt2Cal and Raw2Cal are.
In the Filter a new GetZeroCrossing method has been added.  The main difference
is that it does not have that hard coded threshhold to be a zero crossing anymore.
will find all zero crossing and take the largest one.  This brings it in line with
the other ZeroCrossing methods.

Also fixed a small bug in the two cubic zero finders.  They had zero crossings
defined as cfd[i]>0 && cfd[i+1]<0  where it needs cfd[i]>=0 for the cases
when there is a clocktic at zero