Update to the skeleton histograming program.  It is now compatible with PROOF.  There are two processing scripts, main.C and proof.C.  Main will, as before, process a given run with analoop.cc/h.  Proof.C will open the TChain in a PROOF-lite session and process it with a given number of cores.  Analoop and main.C have been changed slightly to work with PROOF.
Both mian.C and proof.C calls TSelector::GetSelector(fileName)
instead of chain->Process(fileName) because doing so makes PROOF more
happy.

Analoop now saves its histograms internally instead of in main.  This
is done in Terminate method of Analoop.h.

The way the R00TLeSettings objects are handled has changed.  In both
main and proof the R00TLeSettings object FROM THE FIRST FILE is added
to the input list of the TSelector (TSelector::fInputlist).  This way
the settings object is avaiable in both PROOF and not PROOF.