Added new methods to the skeleton histogramer.  MakeHistogram and fill histogram allow one to refer to histograms without having to declare them in the header and new them by hand.  In Slave begin one can type:
MakeHistogram("name",bins,xlow,xhigh)
and then in process

FillHistogram("name",value)

Make histogram stores the histograms in the fOuput of the Selector
in the same way the histograms that are newed explicitly are.

The methods can also take a number instead of a name.  it will then
store things as hTheNumber.