../scripts/MakeProjectionPlotWithIncreasingThresh.C nows takes argument for number of points that you want.
Added AutoHisto Methods to the skeleton Analoop.  Can now say in the process
loop

AutoHisto("name",value,bins,xlow,xhigh);

And it will check to see if the histogram exists.  IF not it makes it
and fills, if it does exists it fills it.  IE can now make histogram
in one line