Changed the LendaPacker to call GetNewFirmwareCFD() to calculate the CFD
Added new methods to LendaFilter:
GetNewFirmwareCFD(), GetNewFirmwareCFDWeight() and CalcualteCFD()

CalcualateCFD is a copy of Chris Prokop's algorithm
GetNewFirmwareCFD is a slight rewrite of CalcualteCFD

Have checked GetNewFirmwareCFD against the Internal Times and they
match on the scale of 0.1 ps or so.  Seems as though the new firmware
makes sense again.

NOTE:  NEVER CHANGE FIRMWARE AGAIN!!!!!!!!!!!