Changed the way the packer handles timing filter settings.  They are now set channel by channel in the map file.  The R00TLeSettings object has been updated to be able to store this information channel by channel.
Remvoed the SetFilter and SetObjectFilter methods in the packer.
Also added new methods to force filters for all Bars, all reference
channels or all channels.  This is need for the TimingAnalyzer
program to be able to set the filters without reading a mapfile.

Added a linkdef statement for the MapInfo Objects, so that one
can use them in the interpreter.

renamed R00TLeSandBox to just Sandbox.