Added new methods to lendaevent.  Has FindBar(int) and FindBar(string) These methods perform linear search on the Bar vector and look for the bar by either BarId or Name.  If it finds them it returns a pointer to it.  If not returns NULL.  Tested this in the interpreter and it seems to behave the way it should.
Also made FindBarSafe(int) and FindBarSaft(string).  Does the same
thing as FindBar except that it returns a copy of the Bar not a
pointer.  Further if the Bar is not in the event it returns an
empty bar object.  This way you don't need to worry about anything
being NULL.  Looks correct in the interpreter

Will probably also add methods to get the underlying Top/bottom
information in a way that won't segfault